python3-requirements.txt
PyYAML
paho-mqtt
icalendar==4.0.3
recurring-ical-events
gevent
